+++ src/riichi/fast_shanten/mod.rs
@@ -1,5 +1,5 @@
 mod chiitoi_classifier;
-mod fast_shanten;
+mod shanten;
 mod kokushi_classifier;
 mod progressive_honor_classifier;
 mod resources;
