+++ src/riichi/fast_shanten/shanten.rs
